2.4.13

+++ drivers/isdn/eicon/linio.c
@@ -1,26 +1,12 @@
-
 /*
- *
  * Copyright (C) Eicon Technology Corporation, 2000.
  *
  * Eicon File Revision :    1.16  
  *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2, or (at your option)
- * any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY OF ANY KIND WHATSOEVER INCLUDING ANY 
- * implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  
- * See the GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
+ * This software may be used and distributed according to the terms
+ * of the GNU General Public License, incorporated herein by reference.
  *
  */
-
 
 #define N_DATA
 
@@ -683,7 +669,7 @@ void UxFree(void *ptr)
 	kfree(ptr);
 }
 
-int UxCardLock(ux_diva_card_t *card)
+long UxCardLock(ux_diva_card_t *card)
 {
 	unsigned long flags;
 
@@ -695,7 +681,7 @@ int UxCardLock(ux_diva_card_t *card)
 	
 }
 
-void UxCardUnlock(ux_diva_card_t *card, int ipl)
+void UxCardUnlock(ux_diva_card_t *card, long ipl)
 {
 	//spin_unlock_irqrestore(&diva_lock, ipl);
